LayoutTests/http/tests/appcache/abort-cache-ondownloading-manifest-404-expected.txt

CONSOLE MESSAGE: Application Cache manifest could not be fetched, because the manifest had a 404 response.
This tests that download process was not aborted after downloading event since manifest was not found.
SUCCESS
